# Use Jemalloc in `forc-lsp` and LSP benchmarks for greater performance.