--- a/forc-plugins/forc-lsp/src/main.rs
+++ b/forc-plugins/forc-lsp/src/main.rs
@@ -2,6 +2,10 @@
 //!
 //! Once installed and available via `PATH`, can be executed via `forc lsp`.
 
+// Use Jemalloc for main binary
+#[global_allocator]
+static GLOBAL: tikv_jemallocator::Jemalloc = tikv_jemallocator::Jemalloc;
+
 use clap::Parser;
 
 #[derive(Debug, Parser)]
@@ -15,6 +19,5 @@ struct App {}
 #[tokio::main]
 async fn main() {
     App::parse();
-
     sway_lsp::start().await
 }