--- a/sway-lsp/benches/bench_main.rs
+++ b/sway-lsp/benches/bench_main.rs
@@ -1,6 +1,10 @@
 mod lsp_benchmarks;
 use criterion::criterion_main;
 
+// Use Jemalloc during benchmarks
+#[global_allocator]
+static GLOBAL: tikv_jemallocator::Jemalloc = tikv_jemallocator::Jemalloc;
+
 criterion_main! {
     lsp_benchmarks::token_map::benches,
     lsp_benchmarks::requests::benches,