--- a/sway-lsp/benches/lsp_benchmarks/mod.rs
+++ b/sway-lsp/benches/lsp_benchmarks/mod.rs
@@ -4,7 +4,6 @@ pub mod token_map;
 
 use lsp_types::Url;
 use std::{path::PathBuf, sync::Arc};
-use sway_core::Engines;
 use sway_lsp::core::session::{self, Session};
 
 pub fn compile_test_project() -> (Url, Arc<Session>) {
@@ -13,8 +12,7 @@ pub fn compile_test_project() -> (Url, Arc<Session>) {
     let uri = Url::from_file_path(benchmark_dir().join("src/main.sw")).unwrap();
     session.handle_open_file(&uri);
     // Compile the project and write the parse result to the session
-    let engines = Engines::default();
-    let parse_result = session::parse_project(&uri, &engines).unwrap();
+    let parse_result = session::parse_project(&uri, &session.engines.read()).unwrap();
     session.write_parse_result(parse_result);
     (uri, Arc::new(session))
 }