## Description

closes #5372

Thanks for the idea @Voxelot , we seem to be getting nice performance wins across all our benchmarks with this change!

| Benchmark                | Avg Duration Before (ms or s) | Avg Duration After (ms or s) | Speed Increase (%) |
|--------------------------|-------------------------------|------------------------------|--------------------|
| compile                  | 5.53213 s                     | 4.64997 s                    | 15.931             |
| traverse                 | 170.646 ms                    | 162.067 ms                   | 5.0397             |
| did_change_with_caching  | 15.332 s                      | 15.4923 s                    | -1.04545           |
| semantic_tokens          | 20.766 ms                     | 15.981 ms                    | 23.053             |
| document_symbol          | 7.9145 ms                     | 5.99825 ms                   | 24.2205            |
| completion               | 49.685 ms                     | 48.5775 ms                   | 2.22975            |
| hover                    | 7.42155 ms                    | 6.23545 ms                   | 15.952             |
| highlight                | 113.75 ms                     | 108.503 ms                   | 4.6195             |
| goto_definition          | 7.99853 ms                    | 6.53863 ms                   | 18.3765            |
| inlay_hints              | 10.6885 ms                    | 8.99667 ms                   | 15.814             |
| prepare_rename           | 7.9312 ms                     | 6.38533 ms                   | 19.4985            |
| rename                   | 119.68 ms                     | 113.803 ms                   | 4.9071             |
| code_action              | 60.1105 ms                    | 55.44 ms                     | 7.7762             |
| code_lens                | 0.365733 ms                   | 0.2341 ms                    | 35.989             |
| on_enter                 | 0.604733 ms                   | 0.3105 ms                    | 48.676             |
| format                   | 51.57 ms                      | 44.6735 ms                   | 13.3765            |
| tokens_for_file          | 10.3933 ms                    | 8.85676 ms                   | 14.723             |
| idents_at_position       | 3.81803 ms                    | 3.1588 ms                    | 17.261             |
| tokens_at_position       | 42.2257 ms                    | 40.9237 ms                   | 3.0873             |
| token_at_position        | 7.67043 ms                    | 6.2267 ms                    | 18.811             |
| parent_decl_at_position  | 42.87 ms                      | 40.8167 ms                   | 4.7931             |